Rank System
evaluate() : This approach will use a rank system and extract the data based on the accuracy of the prediction.

Data will be selected based on a threshold eg: data with >95% accuracy (or based on how many data point the ML can keep). Selected data will be ranked together (or assigned a class). The remaining data will be used to train a different ML. Rinse and repeat until all data is ranked (or until a maximum # of ML is achieve)